backend/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
pdfplumber>=0.11.0
moviepy>=1.0.3
Pillow>=10.0.0
gTTS>=2.5.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
